feat(runtime): share the workload kernel-cmdline assembler with WorkloadRunner
WorkloadRunner::start booted every workload with an empty cmdline,
silently dropping the dm-verity roothash, plan-bound verb-grant
triple, vsock-egress token, and runtime-source-policy tokens the raw
HVF backend already assembled correctly. Extract that assembler into
workload_runner::cmdline (cfg-free, unit-tested with no hypervisor)
and point both the raw HVF backend and the runner's VmBackend::start
at the one implementation, so the two paths can no longer drift apart.
